refactored network layer
