+++ CommonDomain/src/main/java/com/commondomain/network/NetworkConstants.kt
@@ -1,4 +1,4 @@
-package com.commondomain
+package com.commondomain.network
 
 object NetworkConstants {
     const val IMAGE_BASE_URL = "http://image.tmdb.org/t/p/w500"
