+++ CommonDomain/src/main/java/com/commondomain/network/NetworkLauncher.kt
@@ -1,4 +1,4 @@
-package com.commondata.network
+package com.commondomain.network
 
 import kotlinx.coroutines.CoroutineScope
